qemu: Add simple MMU support to UC hooks
Reuse uc->mem_redirect approach to convert address to physical address for ARM.
Add the extra parameter to UC hooks |paddr| which contain a physical address.
Values in |address| and |paddr| can be the same if MMU is not enabled or
physical and virtual addresses are the same.

UC hooks and memory mapping are set _only_ for physical addresses.

Only Python bindings are updated for now to match new hooks signatures.